Phase C-1 cascade: move ResumeData struct to majit-backend::resume_value
After the prior commit converted inherent impl to ResumeDataExt trait
(5d19f3791a), the struct definition itself can move to majit-backend
since the orphan rule no longer applies.

All field types (ResumeValueSource, FrameInfo, VirtualInfo,
PendingFieldInfo) already live in majit-backend.  The
ResumeDataExt trait + impl stay in majit-metainterp because the
methods reference metainterp-specific types (EncodedResumeData,
DecodedResumeLayout, ReconstructedState, MaterializedVirtual etc.).

majit-metainterp::resume re-exports ResumeData for caller
compatibility.  This is the last hop before the ResumeGuardDescr
migration that unblocks backend struct deletion.
